treewide: replace HashMap with IndexMap
This causes both `prune` and `count` to produce deterministic output.

In this change, `FlakeLock::remove_node` calls `IndexMap::swap_remove`,
and that's fine as long as `FlakeNodeVisits` is also deterministic.

If `swap_remove`/`remove_node` is called in random order, the output
will no longer be deterministic. If `shift_remove` were used instead,
the order in which nodes are removed would not matter, it would produce
the same output either way. Currently, `FlakeLock::remove_node`
determinism is dependent upon the order in which it is called.

Closes #1