flake: checks: add formatting-rust check
I'm not sure that `runCommandNoCCLocal` is the right function to use.

On one hand, sending derivations plus source code over the network to
run the formatting checks on another machine would probably be very
wasteful. On the other, keeping this work local means it will probably
be slow to process large codebases on a low-end device, and then
spending the bandwidth may have been worth it.

Also, for some reason in the past, I had decided that `runCommand` was
unsuitable for this type of check, and that it was better to customize
`mkDerivation`'s arguments for the sake of putting the script in
`checkPhase` alone. Since I don't remember my reasoning from that time,
I have opted to just follow what others do, and use `runCommand`.

For future reference, I'm talking about the formatting check for
Hyprnix, found here:
<https://github.com/hyprland-community/hyprnix/blob/30f0e3532cf46080d6b9bcf7b95e3f6d05d8fb43/flake.nix#L70-L85>